比特币支付接口完成

+++ src/main/java/com/jatesun/satinscript/bean/SatinsOrder.java
@@ -21,7 +21,7 @@ public class SatinsOrder implements Serializable {
 
     private Double transFee;
 
-    private Double feeRate;
+    private Integer feeRate;
 
     private Double serviceFee;
